Refactor: MemberService를 인터페이스와 구현체 클래스로 분리
